Change the filter and add the context
Add a context to keep the model and model_id.
Change the domain filter with the new field

[ADD] Add a new field function

This new function allows to get the related documents with
the model and the model_id.
We add this new field because the domain in the view or in the js
doesn't match the wanted result

[UPD] Add missing comma

[ADD] Add unittests